Try fix mul operation parser, revert constant tensor handling

+++ tensorflow/lite/delegates/gpu/common/model_builder.cc
@@ -179,7 +179,7 @@ Status SetAllDimensions<BHWC>(const TfLiteIntArray* dimensions, BHWC* shape) {
 }
 
 DataType ToDataType(TfLiteType type) {
-  std::cout << ">>>DataType: " << type << std::endl;
+  std::cout << ">>>DataType: " << (int)type << std::endl;
   switch (type) {
     case kTfLiteFloat32:
       return DataType::FLOAT32;
